test: 필요 없는 테스트 삭제
