--- backend/baton/src/test/java/touch/baton/fixture/repository/SupporterRepositoryFixture.java
@@ -1,17 +1,0 @@
-package touch.baton.fixture.repository;
-
-import touch.baton.domain.supporter.Supporter;
-import touch.baton.domain.supporter.repository.SupporterRepository;
-
-public class SupporterRepositoryFixture {
-
-    private final SupporterRepository supporterRepository;
-
-    public SupporterRepositoryFixture(final SupporterRepository supporterRepository) {
-        this.supporterRepository = supporterRepository;
-    }
-
-    public Supporter save(final Supporter supporter) {
-        return supporterRepository.save(supporter);
-    }
-}
