--- backend/baton/src/test/java/touch/baton/fixture/repository/TagRepositoryFixture.java
@@ -1,17 +1,0 @@
-package touch.baton.fixture.repository;
-
-import touch.baton.domain.tag.Tag;
-import touch.baton.domain.tag.repository.TagRepository;
-
-public class TagRepositoryFixture {
-
-    private final TagRepository tagRepository;
-
-    public TagRepositoryFixture(final TagRepository tagRepository) {
-        this.tagRepository = tagRepository;
-    }
-
-    public Tag save(final Tag tag) {
-        return tagRepository.save(tag);
-    }
-}
